Fix some mobile issues
aragonUI makes this tough – I really wish it used `em` for margins
instead of `px`.

And though the page looks good in Chrome's inspector in responsive mode,
it does not look good in the iPhone X simulator – there's side scroll,
the drag interactions feel clunky, and editing in general feels awkward.

However, given the tight timeframe, I think this is a decent start. We
need to revisit the design for narrow screens later and consider our
options.